function that performs normalisation of functional data
The function takes 2 inputs: path to a folder with structural data and path to a folder with functional data. The paths should be provided in the Main_fMRI file. The function relies on the naming pattern form SPM, so the realigned images names should start with 'r*' and the field file should start with 'y*'. 

To run it you'll need a file deformation field file that is a result of segmentation and coregistratin. Since we don't have those functions yet, you'll need to create that file first.