funciton that performs normalisation of structural image
The function takes as an argument a path to structural data folder. This path should be provided in Main_fMRI file when executing it. The funciton relies on the naming convention of spm, it requires deformation field file to start with 'y*' and the bias corrected structural image to start with 'm*'.

To run it you need results from coregistration and segmentation, since we don't have scripts for those functions yet, you'll need to get the files first befor running Main file