smoothing.m
This function takes as input the path to the folder with the normalized images (functional). Smoothing is typically applied to normalized functional data rather than structural data in fMRI analysis to enhance the functional signal, reduce spatial noise, and improve the sensitivity of detecting brain activity. It performs spatial smoothing on the data using a Gaussian kernel with a FWHM of 8 mm. The smoothed images are saved with the prefix 's'.